refactor: moved storymap editor to same folder of other pages
