+++ sites/geohub/src/routes/(app)/storymaps/[[id]]/edit/+page.svelte
