+++ sites/geohub/src/routes/(app)/storymaps/[id]/+page.server.ts
@@ -1,5 +1,5 @@
 import type { PageServerLoad } from './$types';
-import { loadStorymap } from './loadStorymap';
+import { loadStorymapById } from '$lib/server/helpers/loadStorymapById';
 
 export const load: PageServerLoad = async (event) => {
 	const { params, parent, fetch } = event;
@@ -7,6 +7,6 @@ export const load: PageServerLoad = async (event) => {
 	const user = session?.user;
 	const id = params.id;
 	const user_email = user?.email as string;
-	const res = await loadStorymap(id, user_email, socialImage, fetch);
+	const res = await loadStorymapById(id, user_email, socialImage, fetch);
 	return res;
 };
