+++ sites/geohub/src/routes/(app)/storymaps/[id]/viewer/+page.server.ts
@@ -1,11 +1,11 @@
 import type { PageServerLoad } from './$types';
-import { loadStorymap } from '../loadStorymap';
+import { loadStorymapById } from '$lib/server/helpers/loadStorymapById';
 
 export const load: PageServerLoad = async ({ fetch, params, parent }) => {
 	const { session, socialImage } = await parent();
 	const user = session?.user;
 	const id = params.id;
 	const user_email = user?.email as string;
-	const res = await loadStorymap(id, user_email, socialImage, fetch);
+	const res = await loadStorymapById(id, user_email, socialImage, fetch);
 	return res;
 };
